Extract scaletestSMTP to separate file
